Add SameSite attribute to the OAuth2 state cookie
* Add a SameSite field to the CookieConfig to allow configuring the
SameSite attribute on the temporary state cookie used by OAuth2
* Set lax mode on the DefaultCookieConfig and DebugOnlyCookieConfig
* Generally, we can't use SameSite strict for the OAuth2 temporary
state cookie. Strict cookies are not sent in the requests to the
callback handler because the redirect chain originated with a redirect
to another domain (e.g. github, google), which is considered the
referrer. Instead, lax mode must be used
* Increase the temporary cookie max age from 1m to 10m to allow users
longer to complete the authorization flow (e.g. login to a provider,
maybe complete 2FA, review permissions, grant permission)

Rel:

* https://www.nogginbox.co.uk/blog/strict-cookies-not-sent-by-request
* https://bugzilla.mozilla.org/show_bug.cgi?id=1453814